requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
docxtpl==0.16.7
python-docx==0.8.11
